Delt spilgenrering op i flere klasser
